Fix database creates requests to lounge
Database create requests to couchdb were working fine, but lounge
apparently requires a content-length header even for empty PUT
requests, so database create requests were failing with a 411 error.
The HTTP spec seems to be a little unclear on whether a content-length
header is required for PUTs... but it seems safer and more correct to
include it, so let's include it for ALL requests.